+++ src/main/java/com/ketheroth/uncrafter/common/inventory/container/UncrafterContainer.java
@@ -9,6 +9,7 @@ import net.minecraft.util.Tuple;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Recipe;
@@ -43,15 +44,13 @@ public class UncrafterContainer extends AbstractContainerMenu implements IUncraf
 		this.outputItems = new OutputHandler(9, this);
 		this.inputItems = new InputHandler(1, this);
 
-		//layout player inventory
-		for (int y = 0; y < 3; y++) {
-			for (int x = 0; x < 9; x++) {
-				this.addSlot(new SlotItemHandler(this.playerInventory, 9 + 9 * y + x, 8 + 18 * x, 84 + 18 * y));
+		//layout input inventory
+		this.addSlot(new SlotItemHandler(this.inputItems, 0, 35, 35) {
+			@Override
+			public boolean mayPickup(Player playerIn) {
+				return !UncrafterContainer.this.outputItems.isExtracting() && super.mayPickup(playerIn);
 			}
-		}
-		for (int i = 0; i < 9; i++) {
-			this.addSlot(new SlotItemHandler(this.playerInventory, i, 8 + 18 * i, 142));
-		}
+		});
 		//layout output inventory
 		for (int y = 0; y < 3; y++) {
 			for (int x = 0; x < 3; x++) {
@@ -63,13 +62,15 @@ public class UncrafterContainer extends AbstractContainerMenu implements IUncraf
 				});
 			}
 		}
-		//layout input inventory
-		this.addSlot(new SlotItemHandler(this.inputItems, 0, 35, 35) {
-			@Override
-			public boolean mayPickup(Player playerIn) {
-				return !UncrafterContainer.this.outputItems.isExtracting() && super.mayPickup(playerIn);
+		//layout player inventory
+		for (int y = 0; y < 3; y++) {
+			for (int x = 0; x < 9; x++) {
+				this.addSlot(new SlotItemHandler(this.playerInventory, 9 + 9 * y + x, 8 + 18 * x, 84 + 18 * y));
 			}
-		});
+		}
+		for (int i = 0; i < 9; i++) {
+			this.addSlot(new SlotItemHandler(this.playerInventory, i, 8 + 18 * i, 142));
+		}
 	}
 
 	@Override
@@ -95,26 +96,42 @@ public class UncrafterContainer extends AbstractContainerMenu implements IUncraf
 
 	@Override
 	public ItemStack quickMoveStack(Player playerIn, int index) {
-		return ItemStack.EMPTY;
-//		ItemStack itemstack = ItemStack.EMPTY;
-//		Slot slot = this.slots.get(index);
-//		if (slot != null && slot.hasItem()) {
-//			ItemStack itemstack1 = slot.getItem();
-//			itemstack = itemstack1.copy();
-//			if (index < playerInventory.getSlots()) {
-//				if (!this.moveItemStackTo(itemstack1, playerInventory.getSlots(), this.slots.size(), true)) {
-//					return ItemStack.EMPTY;
-//				}
-//			} else if (!this.moveItemStackTo(itemstack1, 0, playerInventory.getSlots(), false)) {
-//				return ItemStack.EMPTY;
-//			}
-//			if (itemstack1.isEmpty()) {
-//				slot.set(ItemStack.EMPTY);
-//			} else {
-//				slot.setChanged();
-//			}
-//		}
-//		return itemstack;
+		ItemStack itemstack = ItemStack.EMPTY;
+		Slot slot = this.slots.get(index);
+		if (slot != null && slot.hasItem()) {
+			ItemStack slotStack = slot.getItem();
+			itemstack = slotStack.copy();
+			if (index == 0) {// shift-click in input slot
+				if (!this.moveItemStackTo(slotStack, 10, 46, true)) {
+					return ItemStack.EMPTY;
+				}
+				if (slotStack.isEmpty()) {
+					for (int i = 0; i < 9; i++) {
+						this.outputItems.setStackInSlot(i, ItemStack.EMPTY);
+					}
+				}
+			} else if (1 <= index && index < 10) {// shift-click in output slots
+				// nothing happens
+			} else if (index < 46) {// shift-click in inventory slots
+				if (!slots.get(0).hasItem() || slotStack.is(slots.get(0).getItem().getItem())) {//shift click from inventory
+					if (!this.moveItemStackTo(slotStack, 0, 1, false)) {
+						return ItemStack.EMPTY;
+					}
+					inputItems.fillOutputSlots(itemstack);
+				}
+			}
+
+			if (slotStack.isEmpty()) {
+				slot.set(ItemStack.EMPTY);
+			} else {
+				slot.setChanged();
+			}
+			if (slotStack.getCount() == itemstack.getCount()) {
+				return ItemStack.EMPTY;
+			}
+			slot.onTake(playerIn, slotStack);
+		}
+		return itemstack;
 	}
 
 	@Override