+++ src/main/java/com/ketheroth/uncrafter/common/inventory/container/InputHandler.java
@@ -1,12 +1,20 @@
 package com.ketheroth.uncrafter.common.inventory.container;
 
+import com.ketheroth.uncrafter.common.config.Configuration;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.item.crafting.Recipe;
+import net.minecraft.world.item.crafting.RecipeManager;
+import net.minecraft.world.item.crafting.RecipeType;
+import net.minecraftforge.common.crafting.IShapedRecipe;
 import net.minecraftforge.items.ItemStackHandler;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.Random;
 
 public class InputHandler extends ItemStackHandler {
@@ -50,9 +58,9 @@ public class InputHandler extends ItemStackHandler {
 	public void fillOutputSlots(ItemStack inputStack) {
 		// search ingredient for input item
 		if (!inputStack.is(container.getCache().getA())) {
-			Recipe<?> recipe = UncrafterContainer.searchRecipe(inputStack, container.getRecipeManager());
+			Recipe<?> recipe = searchRecipe(inputStack, container.getRecipeManager());
 			if (recipe != null) {
-				List<ItemStack> list = recipe.getIngredients().stream().collect(ArrayList::new,
+				List<ItemStack> list = convertTo3x3(recipe).stream().collect(ArrayList::new,
 						(accumulator, ingredient) -> accumulator.add(ingredient.isEmpty() ? ItemStack.EMPTY : ingredient.getItems()[new Random().nextInt(ingredient.getItems().length)]),
 						ArrayList::addAll);
 				container.getCache().setA(inputStack.getItem());
@@ -68,5 +76,60 @@ public class InputHandler extends ItemStackHandler {
 			index++;
 		}
 	}
+
+	/**
+	 * Convert a random sized recipe to a 3x3 recipe.
+	 *
+	 * @param recipe the recipe to convert.
+	 * @return the ordered list of ingredients.
+	 */
+	public static List<Ingredient> convertTo3x3(Recipe<?> recipe) {
+		List<Ingredient> ingredients = new ArrayList<>(recipe.getIngredients());
+		if (recipe instanceof IShapedRecipe) {
+			int width = ((IShapedRecipe<?>) recipe).getRecipeWidth();
+			if (width == 2) {
+				ingredients.add(Ingredient.EMPTY);
+				ingredients.add(Ingredient.EMPTY);
+				ingredients.add(Ingredient.EMPTY);
+				ingredients.add(4, Ingredient.EMPTY);
+				ingredients.add(2, Ingredient.EMPTY);
+			} else if (width == 1) {
+				ingredients.add(Ingredient.EMPTY);
+				ingredients.add(Ingredient.EMPTY);
+				ingredients.add(2, Ingredient.EMPTY);
+				ingredients.add(2, Ingredient.EMPTY);
+				ingredients.add(1, Ingredient.EMPTY);
+				ingredients.add(1, Ingredient.EMPTY);
+				ingredients.add(0, Ingredient.EMPTY);
+
+			}
+		}
+		while (ingredients.size() > 9) {
+			ingredients.remove(ingredients.size() - 1);
+		}
+		return ingredients;
+	}
+
+	/**
+	 * Search a recipe having its result item being {@code input}.
+	 *
+	 * @param input         the result item of the recipe.
+	 * @param recipeManager the recipe manager in which the recipe should be searched.
+	 * @return the found recipe, or null if the recipe was not found.
+	 */
+	@Nullable
+	public static Recipe<?> searchRecipe(ItemStack input, RecipeManager recipeManager) {
+		Item inputItem = input.getItem();
+		Optional<Recipe<?>> optionalRecipe = recipeManager.getRecipes().stream()
+				.filter(recipe -> recipe.getType().equals(RecipeType.CRAFTING))
+				.filter(recipe -> !Configuration.BLACKLIST.get().contains(recipe.getId().toString()))
+				.filter(recipe -> !Configuration.IMC_BLACKLIST.contains(recipe.getId().toString()))
+				.filter(recipe -> recipe.canCraftInDimensions(3, 3)
+						&& recipe.getResultItem().getItem() == inputItem
+						&& !recipe.getIngredients().isEmpty())
+				.findAny();
+		return optionalRecipe.orElse(null);
+	}
+
 }